+++ lib/selection_sheet.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'src/selection_sheet.dart';
+export 'src/selection_sheet_form_field.dart';
 export 'src/selection_sheet_theme.dart';
 export 'src/selection_sheet_types.dart';
 export 'src/selection_sheet_view.dart' show SelectionSheetController;